refactor(Tag): Create project tag service
