+++ src/app/teacher/apply-tags-button/apply-tags-button.module.ts
@@ -5,13 +5,13 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatCheckboxModule } from '@angular/material/checkbox';
-import { TagService } from '../../../assets/wise5/services/tagService';
 import { StudentTeacherCommonServicesModule } from '../../student-teacher-common-services.module';
 import { CommonModule } from '@angular/common';
 import { FlexLayoutModule } from '@angular/flex-layout';
 import { FormsModule } from '@angular/forms';
 import { MatDividerModule } from '@angular/material/divider';
 import { ManageTagsDialogComponent } from '../manage-tags-dialog/manage-tags-dialog.component';
+import { ProjectTagService } from '../../../assets/wise5/services/projectTagService';
 
 @NgModule({
   declarations: [ApplyTagsButtonComponent],
@@ -29,6 +29,6 @@ import { ManageTagsDialogComponent } from '../manage-tags-dialog/manage-tags-dia
     MatTooltipModule,
     StudentTeacherCommonServicesModule
   ],
-  providers: [TagService]
+  providers: [ProjectTagService]
 })
 export class ApplyTagsButtonModule {}
