+++ src/app/teacher/create-tag-dialog/create-tag-dialog.component.spec.ts
@@ -1,11 +1,11 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { CreateTagDialogComponent } from './create-tag-dialog.component';
-import { TagService } from '../../../assets/wise5/services/tagService';
 import { MatDialogRef } from '@angular/material/dialog';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { StudentTeacherCommonServicesModule } from '../../student-teacher-common-services.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { ProjectTagService } from '../../../assets/wise5/services/projectTagService';
 
 describe('CreateTagDialogComponent', () => {
   let component: CreateTagDialogComponent;
@@ -20,7 +20,7 @@ describe('CreateTagDialogComponent', () => {
         MatSnackBarModule,
         StudentTeacherCommonServicesModule
       ],
-      providers: [{ provide: MatDialogRef, useValue: { close() {} } }, TagService]
+      providers: [{ provide: MatDialogRef, useValue: { close() {} } }, ProjectTagService]
     });
     fixture = TestBed.createComponent(CreateTagDialogComponent);
     component = fixture.componentInstance;
